Fix JavaScript file, refactor lines in index.js.
This change fixed a bug by adding one line to the index.js file. Specifically, the code was refactored to replace all occurrences of the replaceAll method with a regular expression to remove leading and trailing whitespace and then replace each occurrence with an empty string. This adjustment also added a new line of code to replace all newline and space characters with a single newline character.